refactor: improve code organization and api rest structure
