+++ backend/api/urls.py
@@ -1,9 +1,9 @@
 from django.urls import path
 
-from . import views
+from api.views import health, repo_git, repo_graphql
 
 urlpatterns = [
-    path("ping/", views.ping, name="ping"),
-    path("check-repo/", views.check_repo, name="check_repo"),
-    path("analyze-repo/", views.analyze_repo, name="analyze_repo"),
+    path("ping/", health.ping, name="ping"),
+    path("commits/git/", repo_git.extract_commits, name="extract_commits_from_git"),
+    path("repo/check/", repo_graphql.check_repo, name="check_repo_via_graphql"),
 ]
